+++ library/feature/settings/src/main/kotlin/com/mihaicristiancondrea/android/libs/apptoolkit/feature/settings/ui/general/GeneralSettingsScreen.kt
@@ -88,7 +88,6 @@ fun GeneralSettingsScreen(
         LargeTopAppBarWithScaffold(
             title = title,
             onBackClicked = onBackClicked,
-            actions = { with(contentProvider) { ProvideActions(contentKey = contentKey) } },
             snackbarHostState = snackbarHostState,
             content = content
         )